Start persistent rewrite
