+++ cardano-wallet.cabal
@@ -95,6 +95,9 @@ library
     , mwc-random
     , neat-interpolation
     , optparse-applicative
+    , persistent
+    , persistent-sqlite
+    , persistent-template
     , QuickCheck
     , reflection
     , resourcet
@@ -213,6 +216,8 @@ library
       Cardano.Wallet.Kernel.DB.Spec.Read
       Cardano.Wallet.Kernel.DB.Spec.Update
       Cardano.Wallet.Kernel.DB.Sqlite
+      Cardano.Wallet.Kernel.DB.Sqlite.Beam
+      Cardano.Wallet.Kernel.DB.Sqlite.Persistent
       Cardano.Wallet.Kernel.DB.TxMeta
       Cardano.Wallet.Kernel.DB.TxMeta.Types
       Cardano.Wallet.Kernel.DB.Updates
